+++ app/src/main/java/dev/patrickgold/florisboard/oldsettings/fragments/HomeFragment.kt
@@ -25,10 +25,9 @@ import android.view.ViewGroup
 import dev.patrickgold.florisboard.R
 import dev.patrickgold.florisboard.app.FlorisAppActivity
 import dev.patrickgold.florisboard.databinding.SettingsFragmentHomeBinding
+import dev.patrickgold.florisboard.ime.InputMethodUtils
 import dev.patrickgold.florisboard.oldsettings.SettingsMainActivity
 import dev.patrickgold.florisboard.setup.SetupActivity
-import dev.patrickgold.florisboard.util.checkIfImeIsEnabled
-import dev.patrickgold.florisboard.util.checkIfImeIsSelected
 
 class HomeFragment : SettingsMainActivity.SettingsFragment() {
     private lateinit var binding: SettingsFragmentHomeBinding
@@ -80,8 +79,8 @@ class HomeFragment : SettingsMainActivity.SettingsFragment() {
     }
 
     private fun updateImeIssueCardsVisibilities() {
-        val isImeEnabled = checkIfImeIsEnabled(requireContext())
-        val isImeSelected = checkIfImeIsSelected(requireContext())
+        val isImeEnabled = InputMethodUtils.checkIsFlorisboardEnabled(requireContext())
+        val isImeSelected = InputMethodUtils.checkIsFlorisboardSelected(requireContext())
         binding.imeNotEnabledCard.visibility =
             if (isImeEnabled) {
                 View.GONE
