+++ app/src/main/java/dev/patrickgold/florisboard/setup/EnableImeFragment.kt
@@ -24,8 +24,7 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import dev.patrickgold.florisboard.databinding.SetupFragmentEnableImeBinding
-import dev.patrickgold.florisboard.ime.core.FlorisBoard
-import dev.patrickgold.florisboard.util.checkIfImeIsEnabled
+import dev.patrickgold.florisboard.ime.InputMethodUtils
 
 class EnableImeFragment : Fragment() {
     private lateinit var binding: SetupFragmentEnableImeBinding
@@ -48,7 +47,7 @@ class EnableImeFragment : Fragment() {
 
     override fun onResume() {
         super.onResume()
-        if (checkIfImeIsEnabled(requireContext())) {
+        if (InputMethodUtils.checkIsFlorisboardEnabled(requireContext())) {
             (activity as SetupActivity).changePositiveButtonState(true)
             binding.textAfterEnabled.visibility = View.VISIBLE
         } else {
